Don't use C integer types to make cpplint happy

+++ src/include/tsc.h
@@ -32,11 +32,11 @@ class TimeStampCounter {
  public:
         void start();
         void stop();
-        unsigned long long cycles() const;
+        uint64_t cycles() const;
 
  private:
         union Data {
-            unsigned long long a;
+            uint64_t a;
             unsigned int b[2];
         } m_start, m_end;
 };
@@ -65,7 +65,7 @@ inline void TimeStampCounter::stop() {
 #endif
 }
 
-inline unsigned long long TimeStampCounter::cycles() const {
+inline uint64_t TimeStampCounter::cycles() const {
     return m_end.a - m_start.a;
 }
